rewrite EventsTab in TypeScript
and as a functional component!
